+++ tests/test_jwt.py
@@ -1,4 +1,4 @@
-from django.test import modify_settings
+from django.test import override_settings
 from django.urls import reverse
 from rest_framework.test import APITestCase
 from rest_framework_jwt.settings import api_settings as jwt_api_settings
@@ -7,33 +7,22 @@ from social_core.utils import parse_qs
 from .base import BaseFacebookAPITestCase, BaseTwitterApiTestCase
 
 
-jwt_modify_settings = dict(
-    INSTALLED_APPS={
-        'remove': [
-            'django.contrib.sessions',
-            'rest_framework.authtoken',
-            'knox',
-        ]
-    },
-    MIDDLEWARE_CLASSES={
-        'remove': [
-            'django.contrib.sessions.middleware.SessionMiddleware',
-            'django.contrib.auth.middleware.AuthenticationMiddleware',
-            'django.contrib.messages.middleware.MessageMiddleware',
-        ],
-    }
+jwt_override_settings = dict(
+    INSTALLED_APPS=[
+        'django.contrib.contenttypes',
+        'rest_framework',
+        'social_django',
+        'rest_social_auth',
+        'users',
+    ],
+    MIDDLEWARE=[],
 )
 
 
+@override_settings(**jwt_override_settings)
 class TestSocialAuth1JWT(APITestCase, BaseTwitterApiTestCase):
 
-    @modify_settings(INSTALLED_APPS={'remove': ['rest_framework.authtoken', ]})
     def test_login_social_oauth1_jwt(self):
-        """
-        Currently oauth1 works only if session is enabled.
-        Probably it is possible to make it work without session, but
-        it will be needed to change the logic in python-social-auth.
-        """
         resp = self.client.post(
             reverse('login_social_jwt_user'), data={'provider': 'twitter'})
         self.assertEqual(resp.status_code, 200)
@@ -46,9 +35,9 @@ class TestSocialAuth1JWT(APITestCase, BaseTwitterApiTestCase):
         self.assertEqual(resp.status_code, 200)
 
 
+@override_settings(**jwt_override_settings)
 class TestSocialAuth2JWT(APITestCase, BaseFacebookAPITestCase):
 
-    @modify_settings(**jwt_modify_settings)
     def _check_login_social_jwt_only(self, url, data):
         jwt_decode_handler = jwt_api_settings.JWT_DECODE_HANDLER
         resp = self.client.post(url, data)
@@ -57,7 +46,6 @@ class TestSocialAuth2JWT(APITestCase, BaseFacebookAPITestCase):
         jwt_data = jwt_decode_handler(resp.data['token'])
         self.assertEqual(jwt_data['email'], self.email)
 
-    @modify_settings(**jwt_modify_settings)
     def _check_login_social_jwt_user(self, url, data):
         jwt_decode_handler = jwt_api_settings.JWT_DECODE_HANDLER
         resp = self.client.post(url, data)